requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
PyYAML>=6.0

# Image processing and GeoTIFF support
rasterio>=1.3.0
Pillow>=9.0.0

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Optional: Jupyter notebook support for analysis
jupyter>=1.0.0
ipykernel>=6.0.0